requirements.txt
accelerate==1.11.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.0
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.11.0
asyncpg==0.30.0
attrs==25.4.0
beautifulsoup4==4.14.2
cbor==1.0.0
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
cryptography==46.0.3
dataclasses-json==0.6.7
datasets==4.2.0
dill==0.4.0
distro==1.9.0
fastapi==0.119.0
filelock==3.20.0
FlagEmbedding==1.3.5
frozenlist==1.8.0
fsspec==2025.9.0
greenlet==3.2.4
grpcio==1.75.1
gunicorn==23.0.0
h11==0.16.0
h2==4.3.0
hf-xet==1.1.10
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
huggingface-hub==0.35.3
hyperframe==6.1.0
idna==3.11
ijson==3.4.0.post0
inscriptis==2.6.0
ir_datasets==0.5.11
Jinja2==3.1.6
jiter==0.11.1
joblib==1.5.2
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.27
langchain-community==0.3.31
langchain-core==0.3.79
langchain-ollama==0.3.10
langchain-postgres==0.0.16
langchain-text-splitters==0.3.11
langsmith==0.4.37
lxml==6.0.2
lz4==4.4.4
MarkupSafe==3.0.3
marshmallow==3.26.1
mpmath==1.3.0
multidict==6.7.0
multiprocess==0.70.16
mypy_extensions==1.1.0
networkx==3.5
nltk==3.9.2
numpy==2.3.4
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.3.20
nvidia-nvtx-cu12==12.8.90
ollama==0.6.0
openai==2.6.1
orjson==3.11.3
packaging==25.0
pandas==2.3.3
pdfminer.six==20250506
pdfplumber==0.11.7
peft==0.17.1
pgvector==0.3.6
pillow==12.0.0
portalocker==3.2.0
propcache==0.4.1
protobuf==6.33.0
psutil==7.1.1
psycopg==3.2.12
psycopg-binary==3.2.12
psycopg-pool==3.2.7
pyarrow==21.0.0
pycparser==2.23
pydantic==2.12.2
pydantic-settings==2.11.0
pydantic_core==2.41.4
PyMuPDF==1.26.5
PyPDF2==3.0.1
pypdfium2==5.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.3
qdrant-client==1.15.1
regex==2025.9.18
requests==2.32.5
requests-toolbelt==1.0.0
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.16.2
sentence-transformers==5.1.2
sentencepiece==0.2.1
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
SQLAlchemy==2.0.44
starlette==0.48.0
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tiktoken==0.12.0
tokenizers==0.22.1
torch==2.9.0
tqdm==4.67.1
transformers==4.57.1
trec-car-tools==2.6
triton==3.5.0
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
unlzw3==0.2.3
urllib3==2.5.0
uvicorn==0.37.0
warc3-wet==0.2.5
warc3-wet-clueweb09==0.2.5
xxhash==3.6.0
yarl==1.22.0
zlib-state==0.1.10
zstandard==0.25.0
